+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 UNRELEASED
 ---------------------
 
-* Add the possibility to pass a custom json encoder in call.
+* Add the possibility to pass a custom JSON encoder in call.
 
 0.1.9 (2017-03-24)
 ---------------------